backend/requirements.txt
Flask>=3.0,<4
flask-cors>=4.0,<5
python-dotenv>=1.0,<2
msal>=1.28,<2
requests>=2.31,<3
Flask-Limiter>=3.7,<4
flask-talisman>=1.1,<2
gunicorn>=21.2,<23